requirements.txt
attrs==25.3.0
Django==4.2.23
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
pre-commit==4.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyyaml==6.0
python-decouple==3.8
python-dotenv==1.1.1
pytest==8.4.1
pytest-django==4.5.2
pytest-cov==4.1.0
referencing==0.36.2
rpds-py==0.26.0
ruff==0.12.8
sqlparse==0.5.3
typing-extensions==4.14.1
pytest==8.4.1
